Fixed project pictures
