+++ portfolio-site/gatsby-config.js
@@ -18,7 +18,7 @@ module.exports = {
     {
       resolve: `gatsby-plugin-manifest`,
       options: {
-        name: `gatsby-starter-default`,
+        name: `Antonio Cabrero`,
         short_name: `starter`,
         start_url: `/`,
         background_color: `#663399`,
