--- portfolio-site/src/projectsData.js
@@ -1,29 +1,0 @@
-
-const projectsData = [
-    {
-        name: "Job Tracker",
-        description: "An easy to use application to help keep track of the jobs which you've applied, have followed up and/or interviewd",
-        url: "job.antonioc.dev",
-        repo: "https://github.com/Antonio-Cabrero/Job_tracker/",
-        img: '/src/images/hello.png',
-        lang: "React.js"
-    },
-    {
-        name: "Pomodoro Timer",
-        description: "A tool for increasing productivity and balancing effort and relaxation",
-        url: "pomodoro.antonioc.dev",
-        repo: "https://github.com/Antonio-Cabrero/Pomodoro_Timer",
-        img: "",
-        lang: "JavaScript, HTML and CSS",
-    },
-    {
-        name: "Web App Dashboard",
-        description: "A mockup for a Web app dash board with interactivity",
-        url: "webapp.antonioc.dev",
-        repo: "https://github.com/Antonio-Cabrero/Web-App-Dashboard",
-        img: "",
-        lang: "JavaScript, HTML and CSS"
-    },
-]
-
-export default projectsData
