+++ portfolio-site/src/styles/contact.module.css
@@ -1,5 +1,5 @@
 .Contact_container {
-    padding: 3rem;
+    padding: 0 3rem;
     background: rgba(255, 255, 255, 0.534);
     box-shadow: 0 0 65px rgb(255, 255, 255)
 }
