+++ portfolio-site/src/styles/header.module.css
@@ -1,6 +1,6 @@
 .Header {
     display: block;
-    padding: 2em 0;
+    padding: 2em 0 .5em 0;
     line-height: 1.2;
     flex-direction: column;
     font-size: 1.5em;
