+++ portfolio-site/src/styles/navbar.module.css
@@ -9,8 +9,6 @@
     color: #950740;
     font-weight: 300;
     cursor: pointer;
-    margin-top: 1em;
-    margin-left: .5em;
     text-align: center;
     user-select: none;
     background-color: transparent;
@@ -27,4 +25,17 @@
 .Navbar_btn:focus {
     color: #fff;
     background: #950740;
+}
+
+.Navbar_btn:focus {
+    animation: move .5s ease 1;
+}
+
+@keyframes move {
+    0%   {transform: translate(0,0); background-color: #950740;}
+    25%  {transform: translate(-5px, 0); background-color: #b4097bb6}
+    35%  {transform: translate(5px,0); }
+    50%  {transform: translate(-5px, 0); background-color: #da1e6c}
+    75%  {transform: translate(5px,0); background-color: #910e40}
+    100% {transform: translate(0, 0); background-color: #950740}
 }
